Renamed FeedList to PublicChannelsList, changed icons
